feat: Update Query Location to query with status, fix issue where soft delete actually hard delete in category, category group and location
- Fix issue where soft delete for Category, Category Group and Location is actually soft delete
- Update query Location to query with status
- Fix issue where you can delete a location/category/category group multiple time regardless of status
- Throw Exception when user try to delete a location that still have post and item active
- Update initial data mapping to have 2 default posts and updated initial user avatar

+++ LostAndFound.Infrastructure/Services/Implementations/CategoryGroupService.cs
@@ -120,6 +120,22 @@ namespace LostAndFound.Infrastructure.Services.Implementations
                 throw new EntityWithIDNotFoundException<CategoryGroup>(categoryGroupId);
             }
 
+            if(categoryGroup.IsActive == false)
+            {
+                throw new CategoryGroupAlreadyDisabledException();
+            }
+
+            if (categoryGroup.IsActive == true)
+            {
+                foreach (var category in categoryGroup.Categories)
+                {
+                    if (category.IsActive == true)
+                    {
+                        throw new CategoryGroupStillHaveActiveCategory();
+                    }
+                }
+            }
+
             _categoryGroupRepository.Delete(categoryGroup);
             await _unitOfWork.CommitAsync();
         }
